Rename /admin to /metrics.  Fixes #1234
