+++ server/src/pages/metrics/page.js
