+++ server/src/pages/metrics/view.js
